split useSignAndSendTransaction into 2 hooks
